# src/docs/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"TagTree": {
				"properties": {
					"tagId": {
						"type": "number",
						"format": "double"
					},
					"tagAbbrev": {
						"type": "string"
					},
					"tagGene": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"tagName": {
						"type": "string",
						"nullable": true
					},
					"tagComment": {
						"type": "string",
						"nullable": true
					},
					"tagOrder": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"phenomenId": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"phenomenName": {
						"type": "string"
					},
					"tagEbeneName": {
						"type": "string"
					},
					"tagEbeneId": {
						"type": "number",
						"format": "double"
					},
					"childrenIds": {
						"allOf": [
							{
								"$ref": "#/components/schemas/numberArray"
							}
						],
						"nullable": true
					},
					"parentIds": {
						"allOf": [
							{
								"$ref": "#/components/schemas/numberArray"
							}
						],
						"nullable": true
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/TagTree"
						},
						"type": "array"
					}
				},
				"required": [
					"tagId",
					"tagAbbrev",
					"tagGene",
					"tagName",
					"tagComment",
					"tagOrder",
					"phenomenId",
					"phenomenName",
					"tagEbeneName",
					"tagEbeneId",
					"childrenIds",
					"parentIds",
					"children"
				],
				"type": "object",
				"additionalProperties": true
			},
			"numberArray": {
				"items": {
					"type": "number",
					"format": "double"
				},
				"type": "array",
				"description": "Types generated for queries found in \"src/dao/tag.ts\""
			},
			"ISelectTagsLayersResult": {
				"description": "'SelectTagsLayers' return type",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ISelectOrtTagsResult": {
				"description": "'SelectOrtTags' return type",
				"properties": {
					"numTag": {
						"type": "string",
						"nullable": true
					},
					"tagName": {
						"type": "string"
					},
					"tagLang": {
						"type": "string",
						"nullable": true
					},
					"osmId": {
						"type": "string",
						"nullable": true
					},
					"ortNamelang": {
						"type": "string"
					},
					"lat": {
						"type": "string",
						"nullable": true
					},
					"lon": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"numTag",
					"tagName",
					"tagLang",
					"osmId",
					"ortNamelang",
					"lat",
					"lon"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ISelectPhaenBerResult": {
				"description": "'SelectPhaenBer' return type",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"bezPhaenber": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"bezPhaenber"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ISelectPhaenResult": {
				"description": "'SelectPhaen' return type",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"bezPhaenomen": {
						"type": "string"
					},
					"beschrPhaenomen": {
						"type": "string",
						"nullable": true
					},
					"bezPhaenber": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"bezPhaenomen",
					"beschrPhaenomen",
					"bezPhaenber"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ISelectSinglePhaenResult": {
				"description": "'SelectSinglePhaen' return type",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"bezPhaenomen": {
						"type": "string"
					},
					"beschrPhaenomen": {
						"type": "string",
						"nullable": true
					},
					"bezPhaenber": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"bezPhaenomen",
					"beschrPhaenomen",
					"bezPhaenber"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ISelectAufgabenSetResult": {
				"description": "'SelectAufgabenSet' return type",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"kuerzel": {
						"type": "string"
					},
					"nameAset": {
						"type": "string",
						"nullable": true
					},
					"fokus": {
						"type": "string",
						"nullable": true
					},
					"kommentar": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"kuerzel",
					"nameAset",
					"fokus",
					"kommentar"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ISelectAufgabenResult": {
				"description": "'SelectAufgaben' return type",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"beschreibungAufgabe": {
						"type": "string",
						"nullable": true
					},
					"aufgabenstellung": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"beschreibungAufgabe",
					"aufgabenstellung"
				],
				"type": "object",
				"additionalProperties": true
			},
			"ISelectAufgabenFromSetResult": {
				"description": "'SelectAufgabenFromSet' return type",
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"variante": {
						"type": "number",
						"format": "double"
					},
					"beschreibungAufgabe": {
						"type": "string",
						"nullable": true
					},
					"aufgabenstellung": {
						"type": "string",
						"nullable": true
					},
					"kuerzel": {
						"type": "string"
					},
					"nameAset": {
						"type": "string",
						"nullable": true
					},
					"bezPhaenomen": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"variante",
					"beschreibungAufgabe",
					"aufgabenstellung",
					"kuerzel",
					"nameAset",
					"bezPhaenomen"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "X-API-Key",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "dioe-public-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/tags": {
			"get": {
				"operationId": "GetTags",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TagTree"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/tags/layers": {
			"get": {
				"operationId": "GetTagLayers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ISelectTagsLayersResult"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/tags/ort/{tagId}": {
			"get": {
				"operationId": "GetTagOrte",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ISelectOrtTagsResult"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "tagId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/test/{something}": {
			"post": {
				"operationId": "GetControllerDemo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"test": {
											"type": "string",
											"enum": [
												"ok"
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"test"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "query_param",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "something",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
		"/phaen/ber": {
			"get": {
				"operationId": "GetPhaenBer",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ISelectPhaenBerResult"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/phaen": {
			"get": {
				"operationId": "GetPhaen",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ISelectPhaenResult"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/phaen/{berId}": {
			"get": {
				"operationId": "GetSinglePhaen",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ISelectSinglePhaenResult"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "berId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/aufgaben/sets/{phaenID}": {
			"get": {
				"operationId": "GetAufgabenSets",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ISelectAufgabenSetResult"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "phaenID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/aufgaben/{phaenID}": {
			"get": {
				"operationId": "GetAufgabenPhaen",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ISelectAufgabenResult"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "phaenID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/aufgaben/sets": {
			"get": {
				"operationId": "GetTagOrte",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ISelectAufgabenFromSetResult"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "setID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}